Start teasing apart confusing app dependencies with DI
