+++ flake.nix
@@ -24,9 +24,11 @@
             ps.aiohttp
             ps.aiomqtt
             ps.backoff
+            ps.dependency-injector
             ps.icalendar
             ps.lxml
             ps.mypy
+            # ps.pylint
             ps.pytz
             ps.recurring-ical-events
             ps.types-pillow
